interfaces/jantior,osutil: move janitor to osutil package
